Revert "chore: update release drafter workflow"
